+++ lib/providers/osm_api.dart
@@ -71,15 +71,20 @@ class OsmApiHelper {
   }
 
   Future<List<OsmElement>> elements(Iterable<OsmId> ids) async {
+    const kBatchSize = 500; // elements in a single request
     var client = http.Client();
     try {
       final List<OsmElement> elements = [];
       for (final typ in OsmElementType.values) {
         final typeIds = ids.where((id) => id.type == typ);
-        if (typeIds.isNotEmpty) {
+        for (int i = 0; i < (typeIds.length / kBatchSize).ceil(); i++) {
           final typesName = kOsmElementTypeName[typ]! + 's';
           final url = Uri.https(kOsmEndpoint, '/api/0.6/$typesName', {
-            typesName: typeIds.map((e) => e.ref).join(','),
+            typesName: typeIds
+                .skip(kBatchSize * i)
+                .take(kBatchSize)
+                .map((e) => e.ref)
+                .join(','),
           });
           var request = http.Request('GET', url);
           var response = await client.send(request);